# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meo Mai Moi API",
        "description": "API for Meo Mai Moi, a cat care management platform.",
        "contact": {
            "email": "support@meomaimoi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/invitations": {
            "get": {
                "tags": [
                    "Invitations"
                ],
                "summary": "Get user's sent invitations",
                "description": "Retrieve all invitations sent by the authenticated user.",
                "operationId": "d1dde73b3cddcef30899471105d9bc67",
                "responses": {
                    "200": {
                        "description": "List of sent invitations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "code": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "expires_at": {
                                                        "type": "string",
                                                        "format": "datetime",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    },
                                                    "invitation_url": {
                                                        "type": "string"
                                                    },
                                                    "recipient": {
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invitations"
                ],
                "summary": "Generate a new invitation",
                "description": "Create a new invitation code for the authenticated user.",
                "operationId": "2eb9c68ec032782479236af3a4018dd4",
                "requestBody": {
                    "description": "Optional invitation parameters",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expires_at": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-12-31T23:59:59Z",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invitation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invitation created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "code": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "nullable": true
                                                },
                                                "invitation_url": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Daily invitation limit exceeded"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/pets/{id}/status": {
            "put": {
                "tags": [
                    "Pets"
                ],
                "summary": "Update a pet's status",
                "operationId": "a4e6178a49f13a304fd84c8992cde57b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pet to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "password"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "lost",
                                            "deceased",
                                            "deleted"
                                        ],
                                        "example": "lost"
                                    },
                                    "password": {
                                        "description": "User's current password for confirmation",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pet status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pet": {
                "title": "Pet",
                "required": [
                    "id",
                    "name",
                    "country",
                    "description",
                    "status",
                    "created_by",
                    "pet_type_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Whiskers"
                    },
                    "sex": {
                        "description": "Sex of the pet",
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "not_specified"
                        ],
                        "example": "male"
                    },
                    "birthday": {
                        "description": "Exact birthday (present only when birthday_precision=day). Deprecated: prefer component fields.",
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-01",
                        "nullable": true,
                        "deprecated": true
                    },
                    "birthday_year": {
                        "description": "Birth year when known (year/month/day precision).",
                        "type": "integer",
                        "example": 2020,
                        "nullable": true
                    },
                    "birthday_month": {
                        "description": "Birth month when known (month/day precision).",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "birthday_day": {
                        "description": "Birth day when known (day precision).",
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "birthday_precision": {
                        "description": "Precision level for birthday components.",
                        "type": "string",
                        "enum": [
                            "day",
                            "month",
                            "year",
                            "unknown"
                        ],
                        "example": "month"
                    },
                    "country": {
                        "description": "ISO 3166-1 alpha-2 country code",
                        "type": "string",
                        "example": "VN"
                    },
                    "state": {
                        "type": "string",
                        "example": "Hanoi",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "Hanoi",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "A friendly pet."
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_by": {
                        "description": "ID of user who created this pet",
                        "type": "integer",
                        "example": 5
                    },
                    "pet_type_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}